updated customer retrival

+++ api/src/main/java/org/makila/api/service/postgresql/CustomerService.java
@@ -1,17 +1,19 @@
 package org.makila.api.service.postgresql;
 
+import java.util.List;
+
 import org.makila.api.model.postgresql.Customer;
 import org.makila.api.repository.postgresql.CustomerRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import java.util.List;
+
+import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class CustomerService {
     private final CustomerRepository customerRepository;
 
-    public Customer getCustomerById(Integer id) {
+    public Customer getCustomerById(Long id) {
         return customerRepository.findById(id)
             .orElseThrow(() -> new RuntimeException("Customer not found"));
     }
